Applied HARMONY-3442 [drlvm][winx64] Warnings introduced in VM due to IDATA/UDATA porting to Windows x64.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@522564 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/em/src/EdgeProfileCollector.cpp
@@ -137,7 +137,7 @@ uint32* EdgeMethodProfile::getCounter( uint32 key ) const
     if (val!=key) {
         return NULL;
     }
-    uint32 idx = it - cntMap.begin();
+    uint32 idx = (uint32)(it - cntMap.begin());
     return (uint32*)&counters.front() + idx;
 }
